+++ sdk/storage/storage-file-datalake/src/DataLakeServiceClient.ts
@@ -3,7 +3,7 @@
 
 import "@azure/core-paging";
 
-import { TokenCredential } from "@azure/core-http";
+import { getDefaultProxySettings, isNode, TokenCredential } from "@azure/core-http";
 import { PagedAsyncIterableIterator } from "@azure/core-paging";
 import { BlobServiceClient } from "@azure/storage-blob";
 
@@ -18,9 +18,13 @@ import {
 } from "./models";
 import { Pipeline, StoragePipelineOptions, newPipeline } from "./Pipeline";
 import { StorageClient } from "./StorageClient";
-import { appendToURLPath, appendToURLQuery } from "./utils/utils.common";
+import {
+  appendToURLPath,
+  appendToURLQuery,
+  extractConnectionStringParts
+} from "./utils/utils.common";
 import { createSpan } from "./utils/tracing";
-import { toFileSystemPagedAsyncIterableIterator } from "./transforms";
+import { toDfsEndpointUrl, toFileSystemPagedAsyncIterableIterator } from "./transforms";
 import { ServiceGetUserDelegationKeyOptions, ServiceGetUserDelegationKeyResponse } from "./models";
 import { CanonicalCode } from "@opentelemetry/api";
 import { AccountSASPermissions } from "./sas/AccountSASPermissions";
@@ -47,6 +51,47 @@ export class DataLakeServiceClient extends StorageClient {
    * @memberof DataLakeServiceClient
    */
   private blobServiceClient: BlobServiceClient;
+
+  /**
+   *
+   * Creates an instance of DataLakeServiceClient from connection string.
+   *
+   * @param {string} connectionString Account connection string or a SAS connection string of an Azure storage account.
+   *                                  [ Note - Account connection string can only be used in NODE.JS runtime. ]
+   *                                  Account connection string example -
+   *                                  `DefaultEndpointsProtocol=https;AccountName=myaccount;AccountKey=accountKey;EndpointSuffix=core.windows.net`
+   *                                  SAS connection string example -
+   *                                  `BlobEndpoint=https://myaccount.blob.core.windows.net/;QueueEndpoint=https://myaccount.queue.core.windows.net/;FileEndpoint=https://myaccount.file.core.windows.net/;TableEndpoint=https://myaccount.table.core.windows.net/;SharedAccessSignature=sasString`
+   * @param {StoragePipelineOptions} [options] Optional. Options to configure the HTTP pipeline.
+   * @memberof DataLakeServiceClient
+   */
+  public static fromConnectionString(connectionString: string, options?: StoragePipelineOptions) {
+    options = options || {};
+    const extractedCreds = extractConnectionStringParts(connectionString);
+    if (extractedCreds.kind === "AccountConnString") {
+      if (isNode) {
+        const sharedKeyCredential = new StorageSharedKeyCredential(
+          extractedCreds.accountName!,
+          extractedCreds.accountKey
+        );
+        options.proxyOptions = getDefaultProxySettings(extractedCreds.proxyUri);
+        const pipeline = newPipeline(sharedKeyCredential, options);
+        return new DataLakeServiceClient(toDfsEndpointUrl(extractedCreds.url), pipeline);
+      } else {
+        throw new Error("Account connection string is only supported in Node.js environment");
+      }
+    } else if (extractedCreds.kind === "SASConnString") {
+      const pipeline = newPipeline(new AnonymousCredential(), options);
+      return new DataLakeServiceClient(
+        toDfsEndpointUrl(extractedCreds.url) + "?" + extractedCreds.accountSas,
+        pipeline
+      );
+    } else {
+      throw new Error(
+        "Connection string must be either an Account connection string or a SAS connection string"
+      );
+    }
+  }
 
   /**
    * Creates an instance of DataLakeServiceClient from url.