+++ sdk/storage/storage-file-datalake/test/utils/index.browser.ts
@@ -171,3 +171,8 @@ export function getBrowserFile(name: string, size: number): File {
   file.name = name;
   return file;
 }
+
+export function getSASConnectionStringFromEnvironment(): string {
+  const env = (window as any).__env__;
+  return `BlobEndpoint=https://${env.DFS_ACCOUNT_NAME}.blob.core.windows.net/;QueueEndpoint=https://${env.DFS_ACCOUNT_NAME}.queue.core.windows.net/;FileEndpoint=https://${env.DFS_ACCOUNT_NAME}.file.core.windows.net/;TableEndpoint=https://${env.DFS_ACCOUNT_NAME}.table.core.windows.net/;SharedAccessSignature=${env.DFS_ACCOUNT_SAS}`;
+}